PETS2026/scripts/requirements.txt
numpy>=2.3.2
pandas>=2.3.1
pyarrow>=21.0.0
seaborn>=0.13.2
pyyaml